Load multidocument YAMLs when applying kube resources
