+++ pharos-cluster.gemspec
@@ -40,6 +40,7 @@ Gem::Specification.new do |spec|
   spec.add_runtime_dependency "rouge", "~> 3.1"
   spec.add_runtime_dependency "tty-prompt", "~> 0.16"
   spec.add_runtime_dependency "k8s-client", "~> 0.8.1"
+  spec.add_runtime_dependency "yaml-safe_load_stream", "~> 0.1"
   spec.add_runtime_dependency "excon", "~> 0.62.0"
 
   spec.add_development_dependency "bundler"
