+++ spec/fixtures/stacks/multidoc/resources/multidoc.yml
@@ -1,0 +1,15 @@
+kind: ConfigMap
+apiVersion: v1
+metadata:
+  name: multidoctest
+  namespace: default
+data:
+  doc: 1
+---
+kind: ConfigMap
+apiVersion: v1
+metadata:
+  name: multidoctest2
+  namespace: default
+data:
+  doc: 2
